pyproject.toml
[project]
name = "rib-analyzeur"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdfplumber>=0.11.8",
    "pandas>=2.3.3",
    "python-stdnum>=2.1",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
    "rib-analyzeur",
    "python-dotenv>=1.2.1",
    "google-genai>=1.50.1",
]

[tool.uv.sources]
rib-analyzeur = { workspace = true }
